# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
...
    "/api/customers/{index}/rentals": {
      "get": {
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/rentals/start": {
      "post": {
        "tags": [
          "Rental"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/rentals/stop": {
      "post": {
        "tags": [
          "Rental"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/rentals/paid": {
      "post": {
        "tags": [
          "Rental"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Customer": {
        "required": [
          "birthday",
          "firstName",
          "lastName",
          "street",
          "town",
          "zipCode"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "maxLength": 75,
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date-time"
          },
          "street": {
            "maxLength": 75,
            "type": "string",
            "nullable": true
          },
          "houseNumber": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "town": {
            "maxLength": 75,
            "type": "string",
            "nullable": true
          },
          "bikes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Rental": {
        "required": [
          "begin",
          "bikeID",
          "customerID"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "customerID": {
            "type": "integer",
            "format": "int32"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "bikeID": {
            "type": "integer",
            "format": "int32"
          },
          "bike": {
            "$ref": "#/components/schemas/Bike"
          },
          "begin": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "totalCosts": {
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "paid": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "Bike": {
        "required": [
          "brand",
          "purchaseDate",
          "rentalPriceAddHour",
          "rentalPriceFirstHour"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "brand": {
            "maxLength": 25,
            "type": "string",
            "nullable": true
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "dateOfLastService": {
            "type": "string",
            "format": "date-time"
          },
          "rentalPriceFirstHour": {
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "rentalPriceAddHour": {
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "customers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}